AN-6965 confirmation dialog to time control deletion (#35)
* confirmation popup 
* clock activity theme consistency
* resources cleanup

+++ app/src/main/java/com/chess/clock/fragments/TimeSettingsFragment.java
@@ -42,6 +42,7 @@ import com.chess.clock.entities.AppTheme;
 import com.chess.clock.service.ChessClockLocalService;
 import com.chess.clock.util.MultiSelectionUtil;
 import com.chess.clock.views.StyledButton;
+import com.chess.clock.views.ViewUtils;
 
 import java.util.ArrayList;
 
@@ -357,8 +358,19 @@ public class TimeSettingsFragment extends BaseFragment implements MultiSelection
         if (menuItem.getItemId() == R.id.action_delete) {
             startBtn.performHapticFeedback(HapticFeedbackConstants.LONG_PRESS,
                     HapticFeedbackConstants.FLAG_IGNORE_GLOBAL_SETTING);
-            deleteTimeControls();
-            actionMode.finish(); // Action picked, so close the CAB
+            AlertDialog.Builder builder = new AlertDialog.Builder(getActivity(), R.style.ConfirmationDialogTheme);
+            builder
+                    .setMessage(R.string.delete_custom_time)
+                    .setPositiveButton(R.string.action_delete, (dialog, id) -> {
+                        deleteTimeControls();
+                        actionMode.finish(); // Action picked, so close the CAB
+                    })
+                    .setNegativeButton(R.string.action_keep, (dialog, id) -> {
+                        // Resume the clock
+                    });
+            Dialog dialog = builder.create();
+            ViewUtils.setUpConfirmationPopup(dialog, getResources());
+            dialog.show();
             return true;
         }
         return false;
